requirements.txt
annotated-types==0.6.0
colorama==0.4.6
dnspython==2.6.1
Faker~=23.3.0
mongoengine~=0.28.0
pillow~=10.2.0
pip==24.0
pydantic== 2.6.4
pydantic-settings==2.2.1
pymongo~=3.11.4
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==7.4.2
six==1.16.0
tabulate==0.9.0
typing_extensions~=4.10.0
art~=6.2
cowsay~=6.1
idna~=3.6
httpcore~=1.0.4
httpx~=0.27.0
cryptography~=42.0.5
anyio~=4.3.0
cffi~=1.16.0
attrs~=23.2.0
distro~=1.9.0
wheel~=0.43.0
docutils~=0.20.1
Sphinx~=7.2.6
Jinja2~=3.1.3
filelock~=3.13.1
redis~=5.0.2
bson~=0.5.10
requests~=2.31.0
pytest~=8.1.1
click~=8.1.7
lxml~=5.1.0
certifi~=2024.2.2
email_validator~=2.1.1
pydantic_core~=2.16.3
Deprecated~=1.2.14
wcwidth~=0.2.13
termcolor~=2.4.0
blinker~=1.7.0
PyYAML~=6.0.1
tqdm~=4.66.2
faker==23.3.0